+++ sistema/src/main/java/com/mateus/sistema/domain/ContaPagar.java
@@ -30,7 +30,7 @@ public class ContaPagar extends Conta implements Serializable {
 	public ContaPagar() {
 	}
 
-	public ContaPagar(Integer id, Calendar data, Calendar hora, BigDecimal valor, EstadoPagamento estado,
+	public ContaPagar(Long id, Calendar data, Calendar hora, BigDecimal valor, EstadoPagamento estado,
 			Calendar dataPagamento, Calendar dataVencimento, FormaPagamentoCompra formaPagamentoPedido,
 			ParcelaCompra parcela) {
 		super(id, TipoConta.PAGAR, data, hora, valor, estado, dataPagamento, dataVencimento);
