+++ sistema/src/main/java/com/mateus/sistema/domain/FormaPagamento.java
@@ -18,24 +18,24 @@ public class FormaPagamento implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String descricao;
 	private Integer tipo;
 
 	public FormaPagamento() {
 	}
 
-	public FormaPagamento(Integer id, String descricao, TipoFormaPagamento tipo) {
+	public FormaPagamento(Long id, String descricao, TipoFormaPagamento tipo) {
 		this.id = id;
 		this.descricao = descricao;
 		this.tipo = (tipo == null) ? null : tipo.getCod();
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
