+++ sistema/src/main/java/com/mateus/sistema/domain/FormaPagamentoPedido.java
@@ -19,7 +19,7 @@ public abstract class FormaPagamentoPedido implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	@Column(name = "valor")
 	private BigDecimal valor;
 	@Column(name = "data")
@@ -35,7 +35,7 @@ public abstract class FormaPagamentoPedido implements Serializable {
 	public FormaPagamentoPedido() {
 	}
 
-	public FormaPagamentoPedido(Integer id, FormaPagamento formaPagamento, Calendar data, BigDecimal valor, EstadoPagamento estado) {
+	public FormaPagamentoPedido(Long id, FormaPagamento formaPagamento, Calendar data, BigDecimal valor, EstadoPagamento estado) {
 		this.id = id;
 		this.formaPagamento = formaPagamento;
 		this.data = data;
@@ -44,11 +44,11 @@ public abstract class FormaPagamentoPedido implements Serializable {
 		;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
